# Improve performance of lexers.Get function