The following benchmark test:

```go
func BenchmarkGet(b *testing.B) {
       for i := 0; i < b.N; i++ {
               lexers.Get("go")
       }
}
```

Shows the following results

```
go test -bench BenchmarkGet -run BenchmarkGet -v
```

```
goos: darwin
goarch: amd64
pkg: github.com/alecthomas/chroma/lexers
BenchmarkGet
BenchmarkGet-4   	    9336	    128457 ns/op
PASS
ok  	github.com/alecthomas/chroma/lexers	1.491s
```

Which is not so fast for a function that simply returns a lexer for a language. The reason is that we always try to find a lexer by a filename: https://github.com/alecthomas/chroma/blob/master/lexers/internal/api.go#L54. It has been done due to https://github.com/alecthomas/chroma/issues/94.

@alecthomas what do you think about returning a lexer if we found one by a provided name and only if we didn't find one we'd try to search by file extension? That would dramatically improve the performance:

```
goos: darwin
goarch: amd64
pkg: github.com/alecthomas/chroma/lexers
BenchmarkGet
BenchmarkGet-4   	40540636	        28.1 ns/op
PASS
ok  	github.com/alecthomas/chroma/lexers	1.329s
```

I've included the benchmark test just for demonstrating purposes