@alecthomas 2020-04-25
Seems fairly reasonable, and I can't imagine a situation where it could break existing code really.